requirements.txt
fastapi
uvicorn[standard]
pydantic
torch
numpy
scikit-learn
joblib
pandas
python-dotenv
pinecone
pymupdf 
tiktoken
openai
google-cloud-storage